rename private functions in local_db_management
use _ instead of __ notation
update tests
move test_create_record_directory in func test directory
add test for _check_source_file_exists
